* more singleton methods

+++ lib/generators/templates/migration.rb
@@ -4,7 +4,7 @@ class RatingForMigration < ActiveRecord::Migration
     	t.string  :element_type
     	t.string  :element_attribute
     	t.integer :element_id
-    	t.integer :avg_rating, :default => 0
+    	t.float :avg_rating, :default => 0
     	t.integer :total_rating, :default => 0
     	t.integer :ratings_count, :default => 0
